Move SYCL and Serial impl to separate file
