+++ core/src/decl/Kokkos_Declare_SERIAL.hpp
@@ -20,6 +20,7 @@
 #if defined(KOKKOS_ENABLE_SERIAL)
 #include <Serial/Kokkos_Serial.hpp>
 #include <Serial/Kokkos_Serial_MDRangePolicy.hpp>
+#include <Serial/Kokkos_Serial_ZeroMemset.hpp>
 #endif
 
 #endif
